Update layout component structure and imports
