--- reports/src/lib/hooks.ts
@@ -1,11 +1,0 @@
-import {
-  TypedUseSelectorHook,
-  useDispatch,
-  useSelector,
-  useStore,
-} from "react-redux";
-import type { AppDispatch, AppStore, RootState } from "./store";
-
-export const useAppDispatch = () => useDispatch<AppDispatch>();
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
-export const useAppStore = () => useStore<AppStore>();
